files/requirements.txt
anthropic>=0.40.0
requests>=2.28.0
watchdog>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
